change nwd behaviour - don't move files if not inside `run_and_save`

+++ zntrack/core/base.py
@@ -200,14 +200,12 @@ class NodeBase(zninit.ZnInit):
         nwd = self.__dict__.get("nwd")
         if nwd is None:
             nwd = pathlib.Path("nodes", self.node_name)
-            nwd.mkdir(exist_ok=True, parents=True)
         return nwd
 
     @nwd.setter
     def nwd(self, value: pathlib.Path):
         """Set the node working directory and create the directory."""
         self.__dict__["nwd"] = value
-        value.mkdir(exist_ok=True, parents=True)
 
     @property
     def _descriptor_list(self) -> typing.List[zninit.descriptor.DescriptorTypeT]:
@@ -229,10 +227,16 @@ class Node(NodeBase, metaclass=LoadViaGetItem):
 
     The methods implemented in this class are primarily loading and saving parameters.
     This includes restoring the Node from files and saving results to files after run.
+
+    Attributes
+    ----------
+    _run_and_save: bool, default = False
+        True if inside 'run_and_save'
     """
 
     init_subclass_basecls = NodeBase
     init_descriptors = [zn.params, zn.deps, zn.Method, zn.Nodes, meta.Text] + dvc.options
+    _run_and_save: bool = False
 
     @utils.deprecated(
         reason=(
@@ -419,12 +423,17 @@ class Node(NodeBase, metaclass=LoadViaGetItem):
 
     def run_and_save(self):
         """Main method to run for the actual calculation."""
-        if not self.is_loaded:
-            # Save e.g. the parameters if the Node is not loaded
-            #  this can happen, when using this method outside 'dvc repro'
-            self.save()
-        self.run()
-        self.save(results=True)
+        self._run_and_save = True
+        try:
+            self.nwd.mkdir(exist_ok=True, parents=True)
+            if not self.is_loaded:
+                # Save e.g. the parameters if the Node is not loaded
+                #  this can happen, when using this method outside 'dvc repro'
+                self.save()
+            self.run()
+            self.save(results=True)
+        finally:
+            self._run_and_save = False
 
     # @abc.abstractmethod
     def run(self):
@@ -550,6 +559,7 @@ class Node(NodeBase, metaclass=LoadViaGetItem):
         Use 'dvc status' to check, if the stage needs to be rerun.
 
         """
+        self.nwd.mkdir(parents=True, exist_ok=True)
         _handle_nodes_as_methods(self.zntrack.collect(ZnNodes))
 
         if silent:
@@ -670,36 +680,44 @@ class Node(NodeBase, metaclass=LoadViaGetItem):
         new_ckpt: bool
             True if creating a new checkpoint. False if the checkpoint already existed.
         """
-        utils.update_gitignore(prefix=prefix)
-        nwd = self.nwd.with_name(f"{prefix}_{self.nwd.name}")
+        nwd = self.nwd
+        nwd_new = self.nwd.with_name(f"{prefix}_{self.nwd.name}")
+        nwd_is_new = not nwd_new.exists()
 
-        new_ckpt = not nwd.exists()
+        if self._run_and_save:
+            utils.update_gitignore(prefix=prefix)
 
-        if new_ckpt:
-            log.info(f"Creating new operating directory: {nwd}")
-            log.warning(
-                "Experimental Feature: operating directory is currently not compatible"
-                " with 'dvc exp --temp' or 'dvc exp --queue'"
-            )
-            # TODO add a unique path per node.
-            # TODO check on windows!
-            shutil.copytree(self.nwd, nwd, copy_function=os.link)
+            if nwd_is_new:
+                log.info(f"Creating new operating directory: {nwd_new}")
+                log.warning(
+                    "Experimental Feature: operating directory is currently not"
+                    " compatible with 'dvc exp --temp' or 'dvc exp --queue'"
+                )
+                # TODO add a unique path per node.
+                # TODO check on windows!
+                shutil.copytree(nwd, nwd_new, copy_function=os.link)
+            else:
+                log.info(f"Continuing inside operating directory: {nwd_new}.")
+
+            self.nwd = nwd_new
+            try:
+                yield nwd_is_new
+            except Exception as err:
+                log.warning("Node execution was interrupted.")
+                raise err
+            finally:
+                # Save e.g. `zn.outs` before stopping.
+                self.save(results=True)
+                self.nwd = nwd
+
+            log.info(f"Finished successfully. Moving files from {nwd_new} to {nwd}")
+            shutil.rmtree(nwd)
+            shutil.copytree(nwd_new, nwd, copy_function=os.link)
+            shutil.rmtree(nwd_new)
         else:
-            log.info(f"Continuing inside operating directory: {nwd}.")
-
-        self.nwd = nwd
-        try:
-            yield new_ckpt
-        except Exception as err:
-            log.warning("Node execution was interrupted.")
-            raise err
-        finally:
-            # Save e.g. `zn.outs` before stopping.
-            self.save(results=True)
-
-        nwd = self.nwd.with_name(self.nwd.name.split(f"{prefix}_")[1])
-        log.info(f"Finished successfully. Moving files from {self.nwd} to {nwd}")
-        shutil.rmtree(nwd)
-        shutil.copytree(self.nwd, nwd, copy_function=os.link)
-        shutil.rmtree(self.nwd)
-        self.nwd = nwd
+            # if not inside 'run_and_save' no directory should be created. ?!?!?!
+            self.nwd = nwd_new
+            try:
+                yield nwd_is_new
+            finally:
+                self.nwd = nwd
